+++ src/pages/home/index.tsx
@@ -1,14 +1,13 @@
+import Hero from './Elements/Hero';
+import Header from '../../components/header';
 import * as S from './style'
-import { SectionHero, SectionProject } from './Sections';
-import { Header } from '../../components/index';
 
 const Home : React.FC = ()=>{
 
     return(
         <S.Container>
             <Header/>
-            <SectionHero />
-            <SectionProject/>
+            <Hero/>
         </S.Container>
     )
 }
